@@ -49,7 +49,7 @@ export class IndividualReportComponent implements OnInit {
 
     loadLookups() {
         let employeeParams = new SearchParams();
-        this.employeeService.getEmployees(employeeParams).subscribe((res) => {
+        this.employeeService.getEmployeesLookup(employeeParams).subscribe((res) => {
             this.employeesLookup = res.data;
         });
     }
