@@ -52,9 +52,9 @@ namespace FluentPOS.Shared.Infrastructure.Persistence
         {
             services.AddDbContext<T>(m => m.UseSqlServer(connectionString, e => e.MigrationsAssembly(typeof(T).Assembly.FullName)));
             services.AddHangfire(x => x.UseSqlServerStorage(connectionString));
-            //using var scope = services.BuildServiceProvider().CreateScope();
-            //var dbContext = scope.ServiceProvider.GetRequiredService<T>();
-            //dbContext.Database.Migrate();
+            // using var scope = services.BuildServiceProvider().CreateScope();
+            // var dbContext = scope.ServiceProvider.GetRequiredService<T>();
+            // dbContext.Database.Migrate();
 
             return services;
         }
